+++ example/android/app/src/main/kotlin/com/app/flutter_cart_sample/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.app.flutter_cart_sample
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}