+++ lib/flutter_cart.dart
@@ -1,40 +1,44 @@
 import 'package:flutter/material.dart';
 
 import 'model/cart_model.dart';
+import 'model/cart_response_wrapper.dart';
 
 class FlutterCart {
   static final FlutterCart _instance = FlutterCart._internal();
-  CartItem _cartItem;
+  late CartItem _cartItem;
   // CartItemElement _cartItemElement;
-  List<CartItem> _cartItemList;
+  late List<CartItem> _cartItemList;
   List<CartItem> get cartItem => _cartItemList;
-  List<String> _uuid;
+  late List<String> _uuid;
   bool _filterItemFound = false;
-  var message;
+  late CartResponseWrapper message;
   factory FlutterCart() {
     return _instance;
   }
   FlutterCart._internal() {
-    _cartItemList = new List<CartItem>();
-    _uuid = List<String>();
+    _cartItemList = <CartItem>[];
+    _uuid = [];
   }
 
   /// This method is called when we have to add [productTemp] into cart
   addToCart(
       {@required dynamic productId,
       @required dynamic unitPrice,
-      @required int quantity,
+      String? productName,
+      int quantity = 1,
       dynamic uniqueCheck,
       dynamic productDetailsObject}) {
     _cartItem = new CartItem();
-    _setProductValues(
-        productId, unitPrice, quantity, uniqueCheck, productDetailsObject);
+    _setProductValues(productId, unitPrice, productName, quantity, uniqueCheck,
+        productDetailsObject);
     if (_cartItemList.isEmpty) {
       _cartItem.subTotal =
           double.parse((quantity * unitPrice).toStringAsFixed(2));
       _uuid.add(_cartItem.uuid);
       _cartItemList.add(_cartItem);
-      return {"status": true, "message": _successMessage};
+
+      message = CartResponseWrapper(true, _successMessage, _cartItemList);
+      return message;
     } else {
       _cartItemList.forEach((x) {
         if (x.productId == productId) {
@@ -43,7 +47,14 @@ class FlutterCart {
             if (x.uniqueCheck == uniqueCheck) {
               _filterItemFound = true;
               _updateProductDetails(x, quantity, unitPrice);
-              message = {"status": true, "message": _updateMessage};
+              /* message = {
+                "status": true,
+                "message": _updateMessage,
+                "data": _cartItemList,
+                "length": _cartItemList.length
+              }; */
+              message =
+                  CartResponseWrapper(true, _updateMessage, _cartItemList);
             }
           }
 
@@ -51,7 +62,7 @@ class FlutterCart {
           else {
             _filterItemFound = true;
             _updateProductDetails(x, quantity, unitPrice);
-            message = {"status": true, "message": _successMessage};
+            message = CartResponseWrapper(true, _successMessage, _cartItemList);
           }
         }
       });
@@ -61,8 +72,7 @@ class FlutterCart {
         _uuid.add(_cartItem.uuid);
         _updateProductDetails(_cartItem, quantity, unitPrice);
         _cartItemList.add(_cartItem);
-
-        message = {"status": true, "message": _successMessage};
+        message = CartResponseWrapper(true, _successMessage, _cartItemList);
       }
       _filterItemFound = false;
       return message;
@@ -78,26 +88,39 @@ class FlutterCart {
               .roundToDouble();
     } else {
       _cartItemList.removeAt(index);
-      return {"status": true, "message": _removedMessage};
+      return CartResponseWrapper(true, _removedMessage, _cartItemList);
       // return true;
     }
+    return CartResponseWrapper(true, _removedMessage, _cartItemList);
   }
 
-  deleteItemFromCart(int index, {int specificIndex}) {
+  deleteItemFromCart(int index) {
     for (int i = _cartItemList[index].quantity; i > 0; i--) {
       decrementItemFromCart(index);
     }
+    message = CartResponseWrapper(true, _successMessage, _cartItemList);
+    return message;
   }
 
   deleteAllCart() {
-    _cartItemList = new List<CartItem>();
-    _uuid = List<String>();
+    _cartItemList = <CartItem>[];
+    _uuid = <String>[];
   }
 
-  int findItemIndexFromCart(cartId) {
+  int? findItemIndexFromCart(cartId) {
     for (int i = 0; i < _cartItemList.length; i++) {
       if (_cartItemList[i].productId == cartId) {
         return i;
+      }
+    }
+    return null;
+  }
+
+  CartItem? getSpecificItemFromCart(cartId) {
+    for (int i = 0; i < _cartItemList.length; i++) {
+      if (_cartItemList[i].productId == cartId) {
+        _cartItemList[i].itemCartIndex = i;
+        return _cartItemList[i];
       }
     }
     return null;
@@ -109,15 +132,18 @@ class FlutterCart {
     _cartItemList[index].subTotal =
         (_cartItemList[index].quantity * _cartItemList[index].unitPrice)
             .roundToDouble();
+
+    return CartResponseWrapper(true, _successMessage, _cartItemList);
   }
 
   /// This method is called when we have to [initialize the values in our cart object]
-  void _setProductValues(
-      productId, unitPrice, int quantity, uniqueCheck, productDetailsObject) {
+  void _setProductValues(productId, unitPrice, productName, int quantity,
+      uniqueCheck, productDetailsObject) {
     _cartItem.uuid =
         productId.toString() + "-" + DateTime.now().toIso8601String();
     _cartItem.productId = productId;
     _cartItem.unitPrice = unitPrice;
+    _cartItem.productName = productName;
     _cartItem.quantity = quantity;
     _cartItem.uniqueCheck = uniqueCheck;
     _cartItem.productDetails = productDetailsObject;
@@ -134,7 +160,8 @@ class FlutterCart {
   getCartItemCount() {
     return _cartItemList.length;
   }
- /// This method is called when we have to get the [Total amount]
+
+  /// This method is called when we have to get the [Total amount]
   getTotalAmount() {
     double totalAmount = 0.0;
     _cartItemList.forEach((e) => totalAmount += e.subTotal);