+++ CHANGELOG.md
@@ -11,3 +11,13 @@
 - Product price floating issue fixed.
 
 ## [0.0.5+1]
+
+- Fixed minor changes.
+
+## [0.0.6]
+
+- Flutter Cart updated to null safety.
+- ProductName variable has been added in Cart model.
+- Now you can get the specific product by using getSpecificItemFromCart(productId) named function.
+- Improved the package and fixed some minor issues.
+